pyproject.toml
[project]
name = "ai-code-review-assistant"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "langgraph>=0.2.4",
  "langchain>=0.3.0",
  "langchain-core>=0.3.0",
  "langchain-community>=0.3.0",
  "faiss-cpu>=1.8.0",
  "sentence-transformers>=3.0.1",
  "pydantic>=2.8.2",
  "python-dotenv>=1.0.1",
  "tqdm>=4.66.4",
  "rich>=13.7.1",
  "click>=8.1.7",
  "mcp>=1.1.0"   # Model Context Protocol (python)
]

[tool.setuptools]
packages = ["src"]
